Speed up collision checking by using BVHModel<OBBRSS> instead of BVHModel<RSS>
BVHModel<RSS> was used by DART with the old FCL version. However, the new FCL version is an order of magnitude slower than the old one when using BVHModel<RSS>. I switch to BVHModel<OBBRSS> because that is used by ROS. This brings the DART collision checking back to the old speed.