Removed EXCLUDE_FROM_DEFAULT_BUILD from the apps folder
This target property only affects Visual Studio. It excludes projects from being built when building the solution, when building other projects that depend on the project or when running the project. A project not being rebuilt when it is run is problematic, since the user might run an old version without being aware of it.