Fix BULLET_INCLUDE_DIRS in DART[Core]Config.cmake
The fix from #361 doesn't work if bullet package was found by `find_package`. If bullet is not found, `BULLET_INCLUDE_DIRS` will be set to `BULLET_INCLUDE_DIRS-NOT-FOUND` rather than blank as we assumed in #361. Instead, we set DART_INCLUDE_DIRS in the top CMakeList.txt depending on the condition of that bullet has found, and then use it in DART[Core]Config.cmake.in.