Fix #1223 for the recursive case (#1227)
There's a case where the changes in #1223 fail if dependent components are requested. For example, if you ask for

```
find_package(DART COMPONENTS utils utils-urdf)
```

then the build system can get confused about whether or not the component and its dependencies were properly found. We fix that in this PR by moving the location where the `TRUE` assumption is made.