Fix issue with removing skeletons without shapes
When `subscribeTo` is called with a skeleton in `addSkeleton`, the
skeleton is added to `mSkeletonSources`. However, inside
`removeSkeleton`, `unsubscribeFrom` is not called. Instead,
`removeShapeFrameOf` is called. This is fine for most cases, but if it
fails to remove the skeleton from `mSkeletonSources` if the skeleton
does not have any shapes within it. This causes `mSkeletonSources` to
grow indefinitely. A hazardous side effect of this is that when an
attempt is made to add a new skeleton via `subscribeTo`, the address of
the skeleton might be the same as one of the stale entries in
`mSkeletonSources`. Consequently, the new skeleton's shapes don't get
added to the collision detection engine and very surprising results
follow.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>